yelp-go scrapper code refactor 1
